Fixed critical bug related with index removal

+++ worktest/src/CallBack.hpp
@@ -25,7 +25,7 @@ class CallBack : public King::Action
     public :
     CallBack(const std::function<void()> &func, float seconds);
     ~CallBack();
-    int GetTimeRemaining();
+    float GetTimeRemaining();
     
 protected:
